refactor filters, peer programming
