+++ src/pages/explore/_layout.svelte
@@ -1,24 +1,22 @@
 <script>
-  import { _, locale } from 'svelte-i18n';
   import { onMount, onDestroy } from 'svelte';
   import { fade } from 'svelte/transition';
-  import { clickOutside } from '@/directives';
   import { goto, params } from '@sveltech/routify';
   import { getAllListedGardens } from '@/api/garden';
-  import { geocodeExtensive } from '@/api/mapbox';
   import { allGardens, isFetchingGardens } from '@/stores/garden';
   import Map from '@/components/Map/Map.svelte';
   import Drawer from '@/components/Garden/Drawer.svelte';
   import GardenLayer from '@/components/Map/GardenLayer.svelte';
   import WaymarkedTrails from '@/components/Map/WaymarkedTrails.svelte';
   import routes from '@/routes';
-  import { Progress, LabeledCheckbox, Icon, TextInput, Button, Modal, Tag } from '@/components/UI';
+  import { Progress, LabeledCheckbox, Icon, Button, Tag } from '@/components/UI';
+  import Filter from '@/components/Garden/Filter.svelte';
+  import FilterLocation from '@/components/Garden/FilterLocation.svelte';
   import { getCookie, setCookie } from '@/util';
   import {
     crossIcon,
     cyclistIcon,
     hikerIcon,
-    markerIcon,
     filterIcon,
     bonfireIcon,
     electricityIcon,
@@ -31,8 +29,7 @@
   const fallbackLocation = { longitude: 4.5, latitude: 50.5 };
   let zoom = 7;
 
-  let filteredGardensee;
-  $: filteredGardens = $allGardens;
+  let filteredGardens = null;
 
   $: selectedGarden = $isFetchingGardens ? null : $allGardens[$params.gardenId];
 
@@ -79,98 +76,9 @@
   let showHiking = false;
   let showCycling = false;
 
-  let locationInput = '';
-  //a place format: {latitude: 51.221109, longitude: 4.3997081, place_name: "Antwerpen, Antwerpen, België"}, max 5
-  let places = [];
+  let showFilterModal = false;
 
-  //if there is no text in the inputfield the places array should also be empty
-  $: if (locationInput == '') emptyPlacesAndInput();
-
-  const emptyPlacesAndInput = () => {
-    places = [];
-    locationInput = '';
-  };
-
-  //if locationInput changes then query
-  $: getLocationWithTimeout(locationInput);
-
-  //only query when the user stop typing for 'amount ms' of time (value between 200-1000)
-  let timeout = null;
-  const getLocationWithTimeout = (input) => {
-    if (timeout !== null) {
-      clearTimeout(timeout);
-    }
-    timeout = setTimeout(function () {
-      getlocation(input);
-    }, 250);
-  };
-
-  //if the input is valuable, query the input, biased results based on proximity paramater and locale (language)
-  //if search result is successful, hide no places found warning message
-  let showNoPlacesBool = false;
-  const getlocation = async (string) => {
-    try {
-      if (string.trim() !== '') {
-        const placesReturnedFromGeocode = await geocodeExtensive(
-          string,
-          fallbackLocation.longitude,
-          fallbackLocation.latitude,
-          $locale,
-          5
-        );
-        if (placesReturnedFromGeocode.type == 'succes') {
-          showNoPlacesBool = false;
-          places = placesReturnedFromGeocode.data;
-        } else if (placesReturnedFromGeocode.type == 'error') {
-          places = [];
-          showNoPlacesBool = true;
-          showNoPlaces();
-        }
-      }
-    } catch (e) {
-      places = [];
-      console.log(e);
-    }
-  };
-
-  const displayPlaceName = (place_name) => {
-    const placeParts = place_name.split(',');
-    let placeFormat = '';
-    for (let index = 0; index < placeParts.length - 1; index++) {
-      if (index == placeParts.length - 2) {
-        placeFormat += placeParts[index];
-      } else {
-        placeFormat += placeParts[index] + ', ';
-      }
-    }
-    return placeFormat;
-  };
-
-  //display no places found warning message for 3 seconds, if there is a new warning reset the timer
-  let timeoutNoPlaces = null;
-  const showNoPlaces = () => {
-    if (timeoutNoPlaces !== null) {
-      clearTimeout(timeoutNoPlaces);
-    }
-    timeoutNoPlaces = setTimeout(function () {
-      showNoPlacesBool = false;
-    }, 3000);
-  };
-
-  //go to a place zoom in to level 11 (city) => https://wiki.openstreetmap.org/wiki/Zoom_levels
-  const goToPlace = (long, lat) => {
-    zoom = 11;
-    center = { longitude: long, latitude: lat };
-    emptyPlacesAndInput();
-  };
-
-  //if user drag map, nothing happens, if the user select a garden or click on the map the input and locations are emptied
-  const handleClickOutsidePlaces = (event) => {
-    const { clickEvent } = event.detail;
-    emptyPlacesAndInput();
-  };
-
-  let show = false;
+  let isSearching = false;
 
   let filter = {
     facilities: {},
@@ -180,32 +88,9 @@
     }
   };
 
-  const filterFacilitiesActive = (filter, facilities) => {
-    let filtersActiveArray = [];
-    for (let index = 0; index < facilities.length; index++) {
-      if (filter.facilities[facilities[index].name] == true)
-        filtersActiveArray.push(facilities[index]);
-    }
-    return filtersActiveArray;
+  const activeFacilities = () => {
+    return facilities.filter((facility) => filter.facilities[facility.name] === true);
   };
-
-  /*
-  let filter = {
-    facilities: {
-      bonfire: false,
-      drinkableWater: false,
-      electricity: false,
-      shower: false,
-      tent: false,
-      toilet: false,
-      water: false
-    },
-    capacity: {
-      min: 1,
-      max: 20
-    }
-  };
-*/
 
   //the order is important for displaying the "tags"
   const facilities = [
@@ -217,62 +102,24 @@
     { name: 'water', icon: waterIcon, label: 'Water' },
     { name: 'drinkableWater', icon: waterIcon, label: 'Drinkable water' }
   ];
-
-  const openGardenFilerModal = () => {
-    show = true;
-  };
-
-  const capacityMinReduce = () => {
-    if (filter.capacity.min > 1) filter.capacity.min -= 1;
-  };
-  const capacityMinIncrease = () => {
-    if (filter.capacity.min < 20) filter.capacity.min += 1;
-  };
-
-  function gardenFilterFacilities(garden, index, array) {
-    for (const [key, value] of Object.entries(this)) {
-      if (value && garden.facilities[key] !== value) return false;
-    }
-    return true;
-  }
-
-  function gardenFilterCapacity(garden, index, array) {
-    const value = garden.facilities.capacity;
-    return value >= this.min && value <= this.max;
-  }
-
-  const returnFilteredGardens = (filter, allGardens) => {
-    const filteredGardens = Object.values(allGardens)
-      .filter(gardenFilterFacilities, filter.facilities)
-      .filter(gardenFilterCapacity, filter.capacity);
-
-    let gardens = {};
-    filteredGardens.forEach((garden) => {
-      gardens[garden.id] = { ...garden };
-    });
-    return gardens;
-  };
-
-  const filterGardens = (filter, allGardens) => {
-    filteredGardens = returnFilteredGardens(filter, allGardens);
-    show = false;
-  };
-
-  const uncheckFilterByTag = (facilityName) => {
-    filter.facilities[facilityName] = false;
-
-    filteredGardens = returnFilteredGardens(filter, $allGardens);
-  };
 </script>
 
 <Progress active={$isFetchingGardens} />
+
+<Filter
+  bind:show={showFilterModal}
+  allGardens={$allGardens}
+  bind:filteredGardens
+  {facilities}
+  bind:filter />
+
 <div class="map-section">
   <Map lon={center.longitude} lat={center.latitude} recenterOnUpdate {zoom}>
     {#if !$isFetchingGardens}
       <GardenLayer
         on:garden-click={(e) => selectGarden(e.detail)}
         selectedGardenId={selectedGarden ? selectedGarden.id : null}
-        allGardens={filteredGardens} />
+        allGardens={filteredGardens || $allGardens} />
       <Drawer on:close={closeDrawer} garden={selectedGarden} />
       <WaymarkedTrails {showHiking} {showCycling} />
       <slot />
@@ -321,100 +168,33 @@
 </div>
 
 <div class="filter">
-  <div class="location-filter-column">
-    <div class="location-filter-input">
-      <TextInput
-        icon={markerIcon}
-        type="text"
-        name="location-filter"
-        id="location-filter"
-        placeholder="Search for a city"
-        hideError={true}
-        bind:value={locationInput}
-        autocomplete="off" />
-    </div>
-    {#if places.length >= 1}
-      <div class="location-filter-output">
-        {#each places as place, i}
-          <div
-            on:click={goToPlace(place.longitude, place.latitude)}
-            use:clickOutside
-            on:click-outside={handleClickOutsidePlaces}>
-            <!-- <p>{place.place_name}</p> -->
-            <!-- <p>{place.place_name.split(',')[0]}</p> -->
-            <p>{displayPlaceName(place.place_name)}</p>
-          </div>
-          {#if i != places.length - 1}
-            <hr />
-          {/if}
-        {/each}
-      </div>
-    {:else if showNoPlacesBool}
-      <div class="location-filter-output-error">
-        <p>No places found</p>
-      </div>
-    {:else}
-      <div class="filter-tags">
-        {#each filterFacilitiesActive(filter, facilities) as facility (facility.name)}
-          <Tag
-            name={facility.name}
-            icon={facility.icon}
-            label={facility.label}
-            on:click={() => uncheckFilterByTag(facility.name)} />
-        {/each}
-      </div>
-    {/if}
+  <div class="location-filter">
+    <FilterLocation bind:zoom bind:center bind:isSearching {fallbackLocation} />
   </div>
 
   <div class="garden-filter">
-    <Button type="button" uppercase on:click={openGardenFilerModal}>
+    <Button
+      type="button"
+      uppercase
+      on:click={() => {
+        showFilterModal = true;
+      }}>
       {@html filterIcon}
     </Button>
   </div>
 
+  {#if !isSearching}
+    <div class="filter-tags">
+      {#each activeFacilities() as facility (facility.name)}
+        <Tag
+          name={facility.name}
+          icon={facility.icon}
+          label={facility.label}
+          on:click={() => (filter.facilities[facility.name] = false)} />
+      {/each}
+    </div>
+  {/if}
 </div>
-
-<Modal bind:show maxWidthPX="576" radius="true" center="true" sticky="true">
-  <div slot="title" class="gardenFilterTitleSection">
-    <h2 id="gardenFilterTitle">Filter</h2>
-  </div>
-  <div slot="body" class="gardenFilterBodySection">
-    <hr />
-    <div id="gardenFacilities" class="gardenFilterSection">
-      <h3 class="gardenFilterSubtitle">Garden Facilities</h3>
-      <div class="gardenFilterCheckboxes">
-        {#each facilities as facility (facility.name)}
-          <div class="gardenFilterCheckbox">
-            <LabeledCheckbox {...facility} bind:checked={filter.facilities[facility.name]} />
-          </div>
-        {/each}
-      </div>
-    </div>
-    <hr />
-    <div id="gardenCapacity" class="gardenFilterSection">
-      <h3 class="gardenFilterSubtitle">Garden Capacity</h3>
-      <div class="gardenFilterCapacitySection">
-        <div class="gardenFilterCapacityText">
-          <p>Tent spots available</p>
-        </div>
-        <div class="gardenFilterCapacityModifier">
-          <p>Min.</p>
-          <button on:click={capacityMinReduce}>-</button>
-          <p id="capacity">{filter.capacity.min}</p>
-          <button on:click={capacityMinIncrease}>+</button>
-        </div>
-      </div>
-      <!-- <div class="sliders">
-        <input value="1" min="1" max="20" step="1" type="range" name />
-        <input value="20" min="1" max="20" step="1" type="range" />
-      </div> -->
-    </div>
-  </div>
-
-  <span slot="controls" class="applyGardenFilter">
-    <Button on:click={filterGardens(filter, $allGardens)} medium>Apply Filters</Button>
-  </span>
-</Modal>
 
 <style>
   .map-section {
@@ -448,48 +228,12 @@
     position: absolute;
 
     display: flex;
-    justify-content: flex-end;
+    flex-wrap: wrap;
   }
 
-  .location-filter-column {
-    align-self: stretch;
-    width: 100%;
+  .location-filter {
+    width: calc(100% - 60px);
     margin-right: 1rem;
-  }
-
-  .location-filter-input {
-    background-color: rgba(255, 255, 255);
-    border-radius: 10px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.05);
-    margin-bottom: 1rem;
-  }
-
-  .location-filter-output {
-    background-color: rgba(255, 255, 255);
-    border-radius: 10px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.05);
-    padding: 1rem;
-  }
-
-  .location-filter-output div {
-    cursor: pointer;
-  }
-
-  .location-filter-output hr {
-    height: 1px;
-    background-color: lightgray;
-    border: none;
-  }
-
-  .location-filter-output-error {
-    background-color: rgba(255, 255, 255);
-    border-radius: 10px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.05);
-    padding: 1rem;
-    font-weight: bold;
-    font-style: italic;
-    text-align: center;
-    border: 3px solid var(--color-warning);
   }
 
   .filter-tags {
@@ -510,83 +254,6 @@
     height: 43px;
     margin: 0;
     box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.05);
-  }
-
-  .gardenFilterTitleSection {
-    width: 100%;
-  }
-
-  #gardenFilterTitle {
-    font-weight: bold;
-    font-size: 2rem;
-    text-align: center;
-  }
-
-  .gardenFilterSection {
-    padding: 2.5rem 1.5rem;
-  }
-
-  .gardenFilterSubtitle {
-    font-weight: 600;
-    font-size: 18px;
-    margin-bottom: 1rem;
-    text-align: left;
-  }
-
-  .gardenFilterCheckboxes {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-auto-rows: min-content;
-    grid-row-gap: 1rem;
-  }
-
-  .gardenFilterCapacitySection {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .gardenFilterCapacitySection > div {
-    flex: 0 50%;
-  }
-
-  .gardenFilterCapacityModifier {
-    display: flex;
-    justify-content: center;
-  }
-
-  .gardenFilterCapacityModifier p:first-of-type {
-    margin-left: 1rem;
-  }
-
-  .gardenFilterCapacityModifier > button {
-    margin: 2px 1rem 0 1rem;
-    padding: 0;
-
-    border: 1.5px solid var(--color-green);
-    border-radius: 50%;
-
-    background-color: var(--color-white);
-    cursor: pointer;
-    height: 2rem;
-    width: 2rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .gardenFilterCapacityModifier > button:hover {
-    background-color: var(--color-green);
-    color: var(--color-white);
-  }
-
-  .applyGardenFilter {
-    margin: auto;
-    text-align: center;
-  }
-
-  #capacity {
-    font-weight: bold;
-    border-bottom: 2px solid var(--color-green);
   }
 
   .attribution {
@@ -670,20 +337,6 @@
     .vehicle-notice-wrapper {
       height: 28rem;
     }
-
-    .gardenFilterCapacityText {
-      display: none;
-    }
-
-    .gardenFilterCapacitySection {
-      flex-wrap: wrap;
-      margin: auto;
-      text-align: center;
-    }
-
-    .gardenFilterCapacitySection > div {
-      flex: 0 100%;
-    }
   }
 
   @media screen and (max-width: 500px) {
@@ -726,24 +379,6 @@
       width: 72%;
       left: 50%;
       transform: translateX(-50%);
-    }
-
-    #gardenFilterTitle {
-      font-size: initial;
-    }
-
-    .gardenFilterSection {
-      padding: 1.5rem 0.75rem;
-    }
-
-    .gardenFilterSubtitle {
-      font-weight: 600;
-      font-size: initial;
-      margin-bottom: 0.5rem;
-    }
-
-    .gardenFilterCheckboxes {
-      grid-row-gap: 0.5rem;
     }
   }
 </style>
